additional variable reorg
