+++ variables/__init__.py
@@ -1,6 +1,6 @@
 from variables_parcel import *
 from variables_access import *
-from variables_building import *
 from variables_demographic import *
 from variables_employment import *
 from variables_zone import *
+from variables_building import *
